Updated configs, got equiformer baseline working, testing now

+++ o2s/datasets/radar_dataset.py
@@ -92,7 +92,7 @@ class RadarDataset(MeshXarrayDataset):
             edge_index=non_zero_edges,
             edge_vec=non_zero_edge_vec,
         )
-        response = data.data.permute(2, 0, 1).double()
+        response = data.data.permute(2, 0, 1).float()  # double()
         response = torch.roll(response, response.size(1) // 2, 1)
         # response = torch.view_as_real(data.data)[:, :, :, 0].permute(2, 1, 0).double()
         # response = torch.tensor(torch.view_as_real(data.data)[:, :, :, 0]).permute(
